refactor: group media services by capability
