+++ backend/routes/xhs_routes.py
@@ -18,7 +18,7 @@ from services import (
 )
 from services.database_service import get_db_service
 from services.oss_service import get_oss_service
-from services.video_service import get_video_service
+from services.media import get_video_service
 
 logger = logging.getLogger(__name__)
