+++ backend/scripts/test_toc_planner.py
@@ -19,7 +19,7 @@ def test_full_pipeline():
     from services.llm import init_llm_service
     from services.blog_generator.blog_service import init_blog_service, get_blog_service
     from services.blog_generator.services.search_service import init_search_service
-    from services.image_service import init_image_service
+    from services.media import init_image_service
 
     config = {
         'AI_PROVIDER_FORMAT': os.getenv('AI_PROVIDER_FORMAT', 'openai'),
