+++ backend/services/__init__.py
@@ -13,12 +13,12 @@ _EXPORTS = {
     "init_llm_service": ("services.llm", "init_llm_service"),
     "TransformService": ("services.transform_service", "TransformService"),
     "create_transform_service": ("services.transform_service", "create_transform_service"),
-    "NanoBananaService": ("services.image_service", "NanoBananaService"),
-    "get_image_service": ("services.image_service", "get_image_service"),
-    "init_image_service": ("services.image_service", "init_image_service"),
-    "AspectRatio": ("services.image_service", "AspectRatio"),
-    "ImageSize": ("services.image_service", "ImageSize"),
-    "STORYBOOK_STYLE_PREFIX": ("services.image_service", "STORYBOOK_STYLE_PREFIX"),
+    "NanoBananaService": ("services.media", "NanoBananaService"),
+    "get_image_service": ("services.media", "get_image_service"),
+    "init_image_service": ("services.media", "init_image_service"),
+    "AspectRatio": ("services.media", "AspectRatio"),
+    "ImageSize": ("services.media", "ImageSize"),
+    "STORYBOOK_STYLE_PREFIX": ("services.media", "STORYBOOK_STYLE_PREFIX"),
     "TaskManager": ("services.task_service", "TaskManager"),
     "get_task_manager": ("services.task_service", "get_task_manager"),
     "PipelineService": ("services.pipeline_service", "PipelineService"),
