+++ backend/services/media/__init__.py
@@ -1,0 +1,31 @@
+"""Image and video generation capabilities."""
+
+from .image_service import (
+    AspectRatio,
+    ImageSize,
+    NanoBananaService,
+    STORYBOOK_STYLE_PREFIX,
+    get_image_service,
+    init_image_service,
+)
+from .video_service import (
+    Veo3Service,
+    VideoAspectRatio,
+    get_video_service,
+    get_veo3_service,
+    init_video_service,
+)
+
+__all__ = [
+    "AspectRatio",
+    "ImageSize",
+    "NanoBananaService",
+    "STORYBOOK_STYLE_PREFIX",
+    "Veo3Service",
+    "VideoAspectRatio",
+    "get_image_service",
+    "get_video_service",
+    "get_veo3_service",
+    "init_image_service",
+    "init_video_service",
+]
